frontend: Update create agent form copy
